refs #1866,
tpp no compila en windows 64, hay muchos errores peligroso por lo que lo deshabilitamos por ahora.

la sobrecarga de size_t se queja cuando colisiona con unsigned BINT64



git-svn-id: https://www.tol-project.org/svn/tolp/trunk@6906 be4846ae-f051-483a-831f-b64e7d08527f

+++ tol/init.cpp
@@ -1107,7 +1107,7 @@ CMAKE_SYSTEM == "Linux-3.19.3-100.fc20.x86_64"
               VERSION_BUILD,
               tolBuildDate,
               tolBuildTime,
-              TOL_COMPILER_ID, TOL_BUILD_SYSTEM, sizeof(void*)*8 );
+              TOL_COMPILER_ID, TOL_BUILD_SYSTEM, int(sizeof(void*)*8) ) ;
     init_ = true;
   }
   return aux_;
